backend/requirements.txt
fastapi==0.115.8
openai==1.59.4
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
